refactor: separate websocket connection management from business handlers
