+++ portal/websockets/handlers.py
@@ -1,158 +1,227 @@
 from __future__ import annotations
 
 import json
+import logging
 
-from fastapi import APIRouter, WebSocket, WebSocketDisconnect
+from fastapi import WebSocket
+from sqlalchemy import select
 
-from portal.auth import resolve_booth_role, verify_ws_token
-from portal.booth_identity import make_booth_id as _make_booth_id
-from portal.config import settings
+from portal.auth import can_perform_role
+from portal.booth_identity import parse_booth_id
+from portal.database import get_session as get_db_session
 from portal.globals import booths
-from portal.websockets.manager import (
-    Session,
-    _handle_accept_handoff,
-    _handle_cancel_handoff,
-    _handle_chat,
-    _handle_initiate_handoff,
-    _handle_join,
-    _handle_leave,
-    _handle_set_active,
-    _handle_set_broadcast_unlocked,
-    _handle_update_state,
-    listener_manager,
-    manager,
-    tts_manager,
-)
-
-router = APIRouter()
+from portal.models import DBBooth, Event, Room
+from portal.websockets.manager import Session, manager, listener_manager
 
 
-@router.websocket("/ws/booth/{booth_id}")
-async def ws_booth(websocket: WebSocket, booth_id: str) -> None:
-    try:
-        await verify_ws_token(websocket)
-    except ValueError:
+
+
+
+
+
+async def _handle_join(ws: WebSocket, session: Session, data: dict) -> None:
+    display_name = data.get("display_name", "Interpreter")
+    role = data.get("role", "interpreter")
+    language = data.get("language", "English")
+    channel_id = data.get("channel_id", f"{session.booth_id}-audio")
+    participant_id = data.get("participant_id")
+    if session.granted_role is None:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": "No role assigned for this session."}))
         return
-
-    # Derive granted_role from cookies at connect time — never trust client data.
-    ws_cookies = websocket.cookies
-    ws_session_payload: dict | None = None
-    for cookie_name in ("admin_token", "user_token", "session_token"):
-        raw = ws_cookies.get(cookie_name, "")
-        if not raw:
-            continue
-        try:
-            import jwt as _pyjwt
-
-            ws_session_payload = _pyjwt.decode(
-                raw,
-                settings.effective_jwt_secret,
-                algorithms=["HS256"],
+    if not can_perform_role(session.granted_role, role):
+        await ws.send_text(
+            json.dumps(
+                {
+                    "type": "booth:error",
+                    "message": f"Your assigned role ({session.granted_role}) does not permit joining as {role}.",
+                }
             )
-            break
-        except _pyjwt.InvalidTokenError:
-            continue
+        )
+        return
+    client_event = data.get("event_slug")
+    if client_event is not None:
+        try:
+            await booths.validate_booth_event(session.booth_id, client_event)
+        except PermissionError as exc:
+            await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
+            return
+    room_id = data.get("room_id")
+    if room_id is not None:
+        try:
+            room_id = int(room_id)
+        except (TypeError, ValueError):
+            await ws.send_text(json.dumps({"type": "booth:error", "message": "room_id must be an integer."}))
+            return
+    try:
+        participant, state = await booths.join_participant(
+            booth_id=session.booth_id,
+            display_name=display_name,
+            role=role,
+            language=language,
+            channel_id=channel_id,
+            participant_id=participant_id,
+            room_id=room_id,
+        )
+    except (ValueError, PermissionError) as exc:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
+        return
+    session.participant_id = participant.participant_id
+    session.language = language
+    session.channel_id = channel_id
+    if not state.get("broadcast_unlocked"):
+        try:
+            _slug, _lang = parse_booth_id(session.booth_id)
+            async with get_db_session() as _db:
+                _db_booth = await _db.scalar(
+                    select(DBBooth).join(Event).where(Event.slug == _slug, DBBooth.language_code == _lang)
+                )
+                if _db_booth and _db_booth.broadcast_unlocked:
+                    state = await booths.set_broadcast_unlocked(session.booth_id, True, language, channel_id)
         except Exception as e:
             import logging
 
-            logging.getLogger(__name__).warning(f"Failed to decode token: {e}")
-            continue
-
-    ws_granted_role = await resolve_booth_role(ws_session_payload, booth_id)
-
-    # Validate invite/participant token scope: the token's (event_slug, language_code)
-    # must match the booth being connected to.  This prevents a valid token for booth A
-    # from being used to join booth B.
-    if ws_session_payload is not None and "role" in ws_session_payload:
-        token_event = ws_session_payload.get("event_slug", "")
-        token_lang = ws_session_payload.get("language_code", "")
-        if token_event and token_lang:
-            try:
-                expected_booth_id = _make_booth_id(token_event, token_lang)
-                if expected_booth_id != booth_id:
-                    await websocket.close(code=4003)
-                    return
-            except Exception:
-                await websocket.close(code=4003)
-                return
-
-    await websocket.accept()
-    session = Session(
-        booth_id=booth_id,
-        participant_id=None,
-        language="English",
-        channel_id=f"{booth_id}-audio",
-        granted_role=ws_granted_role,
+            logging.getLogger(__name__).warning(f"Failed to set broadcast unlocked: {e}")
+    await ws.send_text(
+        json.dumps({"type": "booth:joined", "participant_id": participant.participant_id, "state": state})
     )
-    manager.add(websocket, session)
+    await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
 
+
+async def _handle_leave(session: Session) -> None:
+    if not session.participant_id:
+        return
+    state = await booths.leave_participant(
+        session.booth_id, session.participant_id, session.language, session.channel_id
+    )
+    session.participant_id = None
+    await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
+
+
+async def _handle_chat(ws: WebSocket, session: Session, data: dict) -> None:
+    if not session.participant_id:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": "Join the booth before sending messages."}))
+        return
+    body = data.get("body", "")
     try:
-        while True:
-            raw = await websocket.receive_text()
-            try:
-                data = json.loads(raw)
-            except json.JSONDecodeError:
-                await websocket.send_text(json.dumps({"type": "booth:error", "message": "Invalid JSON."}))
-                continue
+        message, state = await booths.add_chat_message(
+            session.booth_id, session.participant_id, body, session.language, session.channel_id
+        )
+    except ValueError as exc:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
+        return
+    await manager.broadcast(session.booth_id, {"type": "booth:chat", "message": message})
+    await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
 
-            msg_type = data.get("type", "")
-            if msg_type == "booth:join":
-                await _handle_join(websocket, session, data)
-            elif msg_type == "booth:leave":
-                await _handle_leave(session)
-            elif msg_type == "booth:chat":
-                await _handle_chat(websocket, session, data)
-            elif msg_type == "booth:set-active":
-                await _handle_set_active(websocket, session, data)
-            elif msg_type == "booth:update-state":
-                await _handle_update_state(websocket, session, data)
-            elif msg_type == "booth:set-broadcast-unlocked":
-                await _handle_set_broadcast_unlocked(websocket, session, data)
-            elif msg_type == "booth:initiate-handoff":
-                await _handle_initiate_handoff(websocket, session, data)
-            elif msg_type == "booth:accept-handoff":
-                await _handle_accept_handoff(websocket, session, data)
-            elif msg_type == "booth:cancel-handoff":
-                await _handle_cancel_handoff(websocket, session, data)
-            else:
-                await websocket.send_text(
-                    json.dumps({"type": "booth:error", "message": f"Unknown message type: {msg_type}"})
-                )
-    except WebSocketDisconnect:
+
+async def _handle_set_active(ws: WebSocket, session: Session, data: dict) -> None:
+    if not session.participant_id:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": "Join the booth first."}))
+        return
+    target_id = data.get("target_id")
+    if not target_id:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": "Missing target_id."}))
+        return
+    snap = await booths.snapshot(session.booth_id, session.language, session.channel_id)
+    previous_active = snap.get("active_interpreter_id")
+    try:
+        state = await booths.set_active_interpreter(
+            session.booth_id, session.participant_id, target_id, session.language, session.channel_id
+        )
+    except (ValueError, PermissionError) as exc:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
+        return
+    if previous_active and previous_active != target_id:
         pass
-    finally:
-        manager.remove(websocket)
-        if session.participant_id:
-            state = await booths.leave_participant(
-                session.booth_id,
-                session.participant_id,
-                session.language,
-                session.channel_id,
+    await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
+
+
+async def _handle_update_state(ws: WebSocket, session: Session, data: dict) -> None:
+    if not session.participant_id:
+        return
+    try:
+        state = await booths.update_participant_state(
+            session.booth_id,
+            session.participant_id,
+            session.language,
+            session.channel_id,
+            mic_active=data.get("mic_active"),
+            ingest_connected=data.get("ingest_connected"),
+            connected=data.get("connected"),
+        )
+    except (ValueError, PermissionError) as exc:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
+        return
+    await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
+
+
+async def _handle_set_broadcast_unlocked(ws: WebSocket, session: Session, data: dict) -> None:
+    if session.granted_role not in ("room_coordinator", "event_owner", "super_admin"):
+        await ws.send_text(
+            json.dumps({"type": "booth:error", "message": "Only Room Coordinators can manage broadcast lock."})
+        )
+        return
+    unlocked = bool(data.get("unlocked"))
+    try:
+        event_slug, language_code = parse_booth_id(session.booth_id)
+        async with get_db_session() as db:
+            stmt = (
+                select(DBBooth)
+                .join(Room)
+                .join(Event)
+                .where(Event.slug == event_slug, DBBooth.language_code == language_code)
             )
-            await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
-
-
-@router.websocket("/ws/captions/{booth_id}")
-async def ws_captions(websocket: WebSocket, booth_id: str) -> None:
-    await websocket.accept()
-    listener_manager.add(websocket, booth_id)
+            result = await db.execute(stmt)
+            db_booth = result.scalar_one_or_none()
+            if db_booth:
+                db_booth.broadcast_unlocked = unlocked
+                await db.commit()
+    except Exception as e:
+        logging.getLogger(__name__).error(f"Error persisting broadcast lock: {e}")
     try:
-        while True:
-            await websocket.receive_text()
-    except WebSocketDisconnect:
-        pass
-    finally:
-        listener_manager.remove(websocket, booth_id)
+        state = await booths.set_broadcast_unlocked(session.booth_id, unlocked, session.language, session.channel_id)
+        await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
+        await listener_manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
+    except Exception as exc:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
 
 
-@router.websocket("/ws/tts/{room_id}/{language_code}")
-async def ws_tts(websocket: WebSocket, room_id: int, language_code: str) -> None:
-    await websocket.accept()
-    tts_manager.add(websocket, room_id, language_code)
+async def _handle_initiate_handoff(ws: WebSocket, session: Session, _data: dict) -> None:
+    if not session.participant_id:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": "Join the booth first."}))
+        return
     try:
-        while True:
-            await websocket.receive_text()
-    except WebSocketDisconnect:
-        pass
-    finally:
-        tts_manager.remove(websocket, room_id, language_code)
+        state = await booths.initiate_handoff(
+            session.booth_id, session.participant_id, session.language, session.channel_id
+        )
+    except (ValueError, PermissionError) as exc:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
+        return
+    await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
+
+
+async def _handle_accept_handoff(ws: WebSocket, session: Session, _data: dict) -> None:
+    if not session.participant_id:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": "Join the booth first."}))
+        return
+    try:
+        state = await booths.accept_handoff(
+            session.booth_id, session.participant_id, session.language, session.channel_id
+        )
+    except (ValueError, PermissionError) as exc:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
+        return
+    await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
+
+
+async def _handle_cancel_handoff(ws: WebSocket, session: Session, _data: dict) -> None:
+    if not session.participant_id:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": "Join the booth first."}))
+        return
+    try:
+        state = await booths.cancel_handoff(
+            session.booth_id, session.participant_id, session.language, session.channel_id
+        )
+    except (ValueError, PermissionError) as exc:
+        await ws.send_text(json.dumps({"type": "booth:error", "message": str(exc)}))
+        return
+    await manager.broadcast(session.booth_id, {"type": "booth:state", "state": state})
